Added grad compute and Updated API

+++ Source/Serrano/operators/nn/common/batchnorm_op.metal
@@ -11,33 +11,33 @@ using namespace metal;
 #include "serrano_misc.h"
 
 typedef struct {
-	short channelPosition; // 0 --> First, 1 --> Last
-	int channels;
-	int inputWidth;
-	int inputHeight;
+    short channelPosition; // 0 --> First, 1 --> Last
+    int channels;
+    int inputWidth;
+    int inputHeight;
     float epsilon;
 } BatchNormInfo;
 
 void kernel batchNorm_inference(constant float* input        [[ buffer(0) ]],
-					device float* output         [[ buffer(1) ]],
-					constant float* mean         [[ buffer(2) ]],
-					constant float* var          [[ buffer(3) ]],
-					constant float* scale        [[ buffer(4) ]],
-					constant float* offset       [[ buffer(5) ]],
-					constant BatchNormInfo& info [[ buffer(6) ]],
-					uint3 thread_id              [[ thread_position_in_grid ]]) {
-	// boundary check
-	if (thread_id.x >= info.inputWidth || thread_id.y >= info.inputHeight) {
-		return;
-	}
-	
-	ChannelPosition channelPosition = getChannelPosition(info.channelPosition);
-	uint inputOutputOffset;
-	if (channelPosition == Last) {
-		inputOutputOffset = thread_id.y * info.inputWidth * info.channels + thread_id.x * info.channels + thread_id.z;
-	} else {
-		inputOutputOffset = thread_id.z * info.inputHeight * info.inputWidth + thread_id.y * info.inputWidth + thread_id.x;
-	}
-	
-	output[inputOutputOffset] = scale[thread_id.z] * (input[inputOutputOffset] - mean[thread_id.z]) / sqrt(var[thread_id.z] + info.epsilon) + offset[thread_id.z];
+                    device float* output         [[ buffer(1) ]],
+                    constant float* mean         [[ buffer(2) ]],
+                    constant float* var          [[ buffer(3) ]],
+                    constant float* scale        [[ buffer(4) ]],
+                    constant float* offset       [[ buffer(5) ]],
+                    constant BatchNormInfo& info [[ buffer(6) ]],
+                    uint3 thread_id              [[ thread_position_in_grid ]]) {
+    // boundary check
+    if (thread_id.x >= uint(info.inputWidth) || thread_id.y >= uint(info.inputHeight)) {
+        return;
+    }
+    
+    ChannelPosition channelPosition = getChannelPosition(info.channelPosition);
+    uint inputOutputOffset;
+    if (channelPosition == Last) {
+        inputOutputOffset = thread_id.y * info.inputWidth * info.channels + thread_id.x * info.channels + thread_id.z;
+    } else {
+        inputOutputOffset = thread_id.z * info.inputHeight * info.inputWidth + thread_id.y * info.inputWidth + thread_id.x;
+    }
+    
+    output[inputOutputOffset] = scale[thread_id.z] * (input[inputOutputOffset] - mean[thread_id.z]) / sqrt(var[thread_id.z] + info.epsilon) + offset[thread_id.z];
 }
